python interface star discrep

+++ tests/cpp/utils.hpp
@@ -3,9 +3,9 @@
 #include <list>
 #include <optional>
 #include <numeric>
-#include <clutchlog/clutchlog.h>
 #include <gtest/gtest.h>
 
+#include "ioh/common/log.hpp"
 #include "ioh/common/file.hpp"
 
 
@@ -13,7 +13,7 @@ inline void expect_vector_eq(const std::vector<int> &x, const std::vector<int> &
 {
     ASSERT_EQ(x.size(), y.size()) << "Vectors x and y are of unequal length";
 
-    for (int i = 0; i < x.size(); ++i)
+    for (size_t i = 0; i < x.size(); ++i)
         EXPECT_EQ(x[i], y[i]) << "Vectors x and y differ at index " << i;
 }
 
@@ -21,7 +21,7 @@ inline void expect_vector_eq(const std::vector<double> &x, const std::vector<dou
 {
     ASSERT_EQ(x.size(), y.size()) << "Vectors x and y are of unequal length";
 
-    for (int i = 0; i < x.size(); ++i)
+    for (size_t i = 0; i < x.size(); ++i)
         EXPECT_NEAR(x[i], y[i], eta) << "Vectors x and y differ at index " << i;
 }
